set code sample text on initial state load
